reword relay proxy intro
This might be a nit, but I don't think we should use the term "microservice" in the context of a single application.  Microservices is, I believe, generally meant to refer to a large scale system architecture in which multiple independently deployable applications are organized around domain bounded contexts.  By this definition it doesn't make much sense to talk about a single application being a "microservice".  I think what we're trying to say here is just that relay proxy is "small".