Update index.mdx (#947)
There are two conflicting callouts at the top of the documents on Single sign-on: the first says that SSO is an Enterprise feature and available to Starter & Pro as an add-on, the second says that SSO & user provisioning are Enterprise features. The first callout is accurate; the second is no longer accurate. Because SSO is an add-on for Starter & Pro, user provisioning with SSO via SAML is also available to those who have the add-on. I moved the second call out to the "User provisioning with SCIM" section and clarified that SCIM is an enterprise-only feature.

The other change was that I added in the user-provisioning-with-scim section was a clarification that the names for the info for setting up a manual SCIM integration may vary by IdP.