Expand description of "average" to mention multiple values (#1775)
This adds some detail to the description of the "average" column in the existing numeric results table, to describe:

* the (existing) behaviour when there are multiple metric values for a given user
* a recently added limit on the number of metric values for a given user

This page will likely be substantially revisited shortly for the Good Solid Experimentation release. Similar information will be needed in the docs for the new product as well. Hopefully having it in here will help as a memento for that work.