Correction to Mean
The mean we display is actually the mean of the posterior, not the average value of the data sent to LaunchDarkly so far. There may be a better way to phrase this - please feel free to alter the wording I suggested, if the experts have better ideas. :)

Also worth noting: this definition exists not just in the docs, but also in the hover-over tooltip of 'Mean' in the experiment results page. Not sure how to ensure that remains consistent, but happy to help if you can point me in the right direction!