(maint) Reword Zapier pricing tier callout
The Zapier integration itself does not limit API calls - we only
restrict Starter accounts via a flag for the integration. This updates
the integration docs to clarify that the Zapier integration can be used
with both a Pro and Enterprise account, but some functionality is
limited to the Enterprise tier based on API usage (for example the Teams
API is Enterprise-only).