Add missing docs for click and pageview events (#2150)
We noticed that the data export schema reference mentions custom events but not click or pageview events, even though we do export all three. (They were documented in the individual schema references for mParticle and Segment, but not in the top-level schema reference page covering all other destinations.) This adds them.