Add callout redirecting customers and integrators to metric import API (#2171)
A fair number of people - including customers, LD SEs, and random third party data vendors - have discovered our internal event ingest endpoint, intended only for use by SDKs, via the SDK Contributor's Guide, and built their own metric import mechanisms on top of it. We want to discourage this now that we have a metric import API specifically for that use case.

This adds a callout to that section of the contributor's guide to redirect those people to the metric import API docs added in #2139.