Cross-link Segment Experimentation integration with Metric Import (#2216)
As suggested by @mmrj, since these both address the same use case. This aims to improve discoverability and clarify the concept of both features.

Also moved the metric import docs into the "Creating experiments" subtree, for consistency with the Segment integration docs, and added a redirect so the old URL doesn't break.