[SC-182052] Update Slack integration documentation with changes
This updates the Slack integration documentation to follow new workflows
and include updated information and screenshots based on recent changes
to the app. That includes using the new `subscribe` and `list` commands, updating
the descriptions of Slack's functionality to emphasize approval
interactions and de-emphasize toggling flags, and removing references to
`filters` and `flags`.