Metric Import: document User-Agent header, and emphasize array (#2780)
A couple of changes to the EAP docs:

(sc-183309) document the UserAgent header we would like partners to send with their events, consistent with other guides for integration partners. (It'd be nice if customers sent this too, although less essential.)
(sc-185662) call more attention to the requirement for the request body to be a JSON array. A partner tried sending a single top-level JSON object, probably influenced by the "example object" snippet which makes it look like that is valid, but it isn't. Especially while we don't validate the body, this makes it worth repeating ourselves a bit.