update endOfLife.tsx to account for releases on previous major versions (#3020)
* update endOfLife.tsx to account for releases on previous major versions

The current script would assign (next release + 1 year) EOL dates to all releases. This is undesirable in the
scenario where we put out a minor release on a previous major version: that would cause all releases prior to it
to have EOL dates relative to it, which would extend the support window out beyond the *current* major version.

This change caps the EOL date at min(next version + 1 year, next major + 1 year). The effect is that we can put out
minors on older major branches, and their EOL dates will have an upper bound of the major's release + 1 year.